crwal4ai/admin_remix/requirements.txt
# Flask Admin Panel Dependencies
Flask>=3.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# Crawler Dependencies (admin_remix içinde crawler çalıştırıyor)
beautifulsoup4>=4.12.0
selenium>=4.15.0
